Refactor and add norwegian text
